refactor: member 엔티티 관련 enum 클래스 위치 이동
