+++ src/main/java/com/dnd/bbok/member/domain/OAuth2Provider.java
@@ -1,4 +1,4 @@
-package com.dnd.bbok.member.adapter.out.persistence.entity;
+package com.dnd.bbok.member.domain;
 
 import lombok.Getter;
